refactor : main 메서드 이동 (Main.cass -> LottoMachine.class)
