+++ src/main/java/lotto/controller/LottoMachine.java
@@ -34,4 +34,13 @@ public class LottoMachine {
         LottoResult result = userLottoPaper.getResultCompareWith(winnerBalls);
         Output.printStatisticsToUser(userMoney, result);
     }
+
+    public static void main(String[] args) {
+        LottoMachine machine = new LottoMachine();
+        machine.inputPrice();
+        machine.generateAuto();
+        machine.outputAboutPurchasedLotto();
+        machine.inputWinnerInformation();
+        machine.outputStatisticsAboutPurchasedLotto();
+    }
 }
